♻️ Refactor: Review 반영 리팩토링
