+++ src/api/team-mood-tracker/end-point.constants.ts
@@ -3,3 +3,13 @@ export const MOOD_TRACKER_API_ENDPOINTS = {
   RESPONSES: (moodTrackerHashedId: string) => `/mood-trackers/${moodTrackerHashedId}/responses`,
   MODIFY_TITLE: (moodTrackerHashedId: string) => `/mood-trackers/${moodTrackerHashedId}`,
 } as const;
+
+export const TEAM_MOOD_TRACKER_PAGE_ROUTES = {
+  /**
+   * 팀 분위기 트래커 다운로드 페이지
+   * @param workspaceId 워크스페이스 ID
+   * @param moodTrackerHashedId 팀 무드 트래커 해시 ID
+   */
+  DOWNLOAD: (workspaceId: string, moodTrackerHashedId: string) =>
+    `/workspace/${workspaceId}/team-mood-tracker/${moodTrackerHashedId}/download`,
+} as const;
